// src/app/api/models/trade-cards-models-legal-entities-legal-entity-brief-model.ts
/* tslint:disable */

/**
 * Строка таблицы списка юр. лиц
 */
export interface TradeCardsModelsLegalEntitiesLegalEntityBriefModel {

  /**
   * Дата архивирования
   */
  archiveDate?: string;

  /**
   * Контактный Email
   */
  contactEmail?: string;

  /**
   * ФИО контактного лица
   */
  contactPersonName?: string;

  /**
   * Контактный телефон
   */
  contactPhone?: string;

  /**
   * Дата создания
   */
  createDate?: string;

  /**
   * Идентификатор
   */
  id?: string;

  /**
   * Наименование
   */
  name?: string;

  /**
   * Количество пунктов самовывоза
   */
  pickupPointsCount?: number;

  /**
   * Количество терминалов
   */
  terminalsCount?: number;
}